Move helpers to utility classes
