+++ src/integrationTest/java/uk/gov/hmcts/reform/bulkscan/orchestrator/controllers/AttachExceptionRecordToExistingCaseTest.kt
@@ -38,7 +38,7 @@ import uk.gov.hmcts.reform.bulkscan.orchestrator.config.Environment.CASE_TYPE_BU
 import uk.gov.hmcts.reform.bulkscan.orchestrator.config.Environment.JURISDICTION
 import uk.gov.hmcts.reform.bulkscan.orchestrator.config.Environment.USER_ID
 import uk.gov.hmcts.reform.bulkscan.orchestrator.config.IntegrationTest
-import uk.gov.hmcts.reform.bulkscan.orchestrator.config.PortWaiter.waitFor
+import uk.gov.hmcts.reform.bulkscan.orchestrator.util.PortWaiter.waitFor
 import uk.gov.hmcts.reform.ccd.client.model.CallbackRequest
 import uk.gov.hmcts.reform.ccd.client.model.CallbackRequest.CallbackRequestBuilder
 import uk.gov.hmcts.reform.ccd.client.model.CaseDetails
