+++ src/integrationTest/java/uk/gov/hmcts/reform/bulkscan/orchestrator/controllers/SupplementaryEvidenceCreatorTest.kt
@@ -16,6 +16,7 @@ import org.springframework.test.context.junit.jupiter.SpringExtension
 import uk.gov.hmcts.reform.bulkscan.orchestrator.config.Environment.CASE_EVENT_URL
 import uk.gov.hmcts.reform.bulkscan.orchestrator.config.Environment.GET_CASE_URL
 import uk.gov.hmcts.reform.bulkscan.orchestrator.config.IntegrationTest
+import uk.gov.hmcts.reform.bulkscan.orchestrator.util.MessageSender
 import java.io.File
 import java.util.concurrent.TimeUnit
